src/requirements.txt
black
crispy-tailwind
django == 5.2.1
django_browser_reload
django-admin-honeypot-updated-2021
django-cors-headers
django-crispy-forms
django-leaflet
django-otp
django-phonenumber-field
django-tailwind
djangorestframework == 3.15.2
djangorestframework_gis
drf-spectacular
isort
markdown
phonenumbers
pre-commit
psycopg2-binary
pytest == 8.3.5
pytest-black
pytest-cov
pytest-django
pytest-sugar
python-decouple
qrcode
requests
tqdm
